sysdb: make new_subdomain() public
Resolves: https://github.com/SSSD/sssd/issues/5151
